+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # :package: 0.1.0
 
+01. :gift: Added exception `dev.orne.beans.UnrecognizedIdentityTokenException`
 01. :gift: Added class `dev.orne.beans.IdentityTokenFormatter`
 01. :gift: Added annotation `dev.orne.beans.ValidIdentityToken`
 01. :gift: Added annotation `dev.orne.beans.ValidIdentityTokenPrefix`
 01. :gift: Added interface `dev.orne.beans.Identity`
 01. :gift: Added class `dev.orne.beans.TokenIdentity`
-01. :gift: Added exception `dev.orne.beans.UnrecognizedIdentityTokenException`
+01. :gift: Added class `dev.orne.beans.AbstractIdentity`
 01. :gift: Added annotation `dev.orne.beans.IdentityTokenResolver`
 01. :gift: Added class `dev.orne.beans.IdentityResolver`
 01. :gift: Added interface `dev.orne.beans.IdentityBean`